[#1] Merge util package into prelude
