+++ prelude/cardano-prelude.cabal
@@ -12,10 +12,47 @@ build-type:          Simple
 extra-source-files:  ChangeLog.md, README.md
 cabal-version:       >=1.10
 
+flag development
+    description: Disable `-Werror`
+    default: False
+    manual: True
+
 library
   hs-source-dirs:      src
-  exposed-modules:     Cardano.Prelude
-  build-depends:       base >=4.11 && <4.12
+
+  exposed-modules:
+                       Cardano.Prelude
+
+  other-modules:
+                       Cardano.Prelude.Base
+                       Cardano.Prelude.Error
+                       Cardano.Prelude.Formatting
+                       Cardano.Prelude.Json.Canonical
+                       Cardano.Prelude.Json.Parse
+                       Cardano.Prelude.Orphans
+                       Cardano.Prelude.Parse
+
+  build-depends:       base
+                     , aeson
+                     , base16-bytestring
+                     , bytestring
+                     , canonical-json
+                     , cborg
+                     , containers
+                     , formatting
+                     , hashable
+                     , mtl
+                     , tagged
+                     , text
                      , universum
+
   default-language:    Haskell2010
   default-extensions:  NoImplicitPrelude
+
+  ghc-options:         -Weverything
+                       -fno-warn-missing-import-lists
+                       -fno-warn-unsafe
+                       -fno-warn-safe
+
+  if (!flag(development))
+    ghc-options:         -Werror
