+++ prelude/src/Cardano/Prelude.hs
@@ -1,5 +1,11 @@
 module Cardano.Prelude
-       ( module Universum
+       ( module X
        ) where
 
-import           Universum
+import           Cardano.Prelude.Base as X
+import           Cardano.Prelude.Error as X
+import           Cardano.Prelude.Formatting as X
+import           Cardano.Prelude.Json.Canonical as X
+import           Cardano.Prelude.Json.Parse as X
+import           Cardano.Prelude.Orphans ()
+import           Cardano.Prelude.Parse as X
