+++ prelude/src/Cardano/Prelude/Base.hs
@@ -1,0 +1,5 @@
+module Cardano.Prelude.Base
+       ( module Universum
+       ) where
+
+import           Universum
