+++ prelude/test/Test/Cardano/Prelude.hs
@@ -1,0 +1,11 @@
+module Test.Cardano.Prelude
+       ( module X
+       ) where
+
+import           Test.Cardano.Prelude.Base16 as X
+import           Test.Cardano.Prelude.Gen as X
+import           Test.Cardano.Prelude.Golden as X
+import           Test.Cardano.Prelude.Orphans ()
+import           Test.Cardano.Prelude.QuickCheck.Arbitrary as X
+import           Test.Cardano.Prelude.QuickCheck.Property as X
+import           Test.Cardano.Prelude.Tripping as X
