+++ prelude/test/Test/Cardano/Prelude/Gen.hs
@@ -1,9 +1,9 @@
-module Test.Cardano.Util.Gen
+-- | Hedgehog generators for commonly used types
+
+module Test.Cardano.Prelude.Gen
        ( genBytes
        , genUTF8Byte
        , gen32Bytes
-       , genMillisecond
-       , genMicrosecond
        , genWord32
        , genWord16
        , genNatural
@@ -11,11 +11,10 @@ module Test.Cardano.Util.Gen
 
 import           Cardano.Prelude
 
-import           Data.Time.Units (Microsecond, Millisecond, fromMicroseconds)
-
-import           Hedgehog
+import           Hedgehog (Gen)
 import qualified Hedgehog.Gen as Gen
 import qualified Hedgehog.Range as Range
+
 
 genBytes :: Int -> Gen ByteString
 genBytes n = Gen.bytes (Range.singleton n)
@@ -25,12 +24,6 @@ genUTF8Byte = Gen.utf8 (Range.constant 0 64) Gen.alphaNum
 
 gen32Bytes :: Gen ByteString
 gen32Bytes = genBytes 32
-
-genMillisecond :: Gen Millisecond
-genMillisecond = fromMicroseconds <$> Gen.integral (Range.constant 0 1000000)
-
-genMicrosecond :: Gen Microsecond
-genMicrosecond = fromMicroseconds <$> Gen.integral (Range.constant 0 1000000)
 
 genWord32 :: Gen Word32
 genWord32 = Gen.word32 Range.constantBounded
