+++ prelude/test/cardano-prelude-test.cabal
@@ -1,4 +1,4 @@
-name:                cardano-util-test
+name:                cardano-prelude-test
 version:             0.1.0.0
 synopsis:            Utility types and functions for testing Cardano
 description:         Utility types and functions for testing Cardano
@@ -11,15 +11,23 @@ category:            Currency
 build-type:          Simple
 cabal-version:       >=1.10
 
+flag development
+    description: Disable `-Werror`
+    default: False
+    manual: True
+
 library
   exposed-modules:
-                       Test.Cardano.Util.Base16
-                       Test.Cardano.Util.Gen
-                       Test.Cardano.Util.Golden
-                       Test.Cardano.Util.Orphans
-                       Test.Cardano.Util.QuickCheck.Arbitrary
-                       Test.Cardano.Util.QuickCheck.Property
-                       Test.Cardano.Util.Tripping
+                       Test.Cardano.Prelude
+
+  other-modules:
+                       Test.Cardano.Prelude.Base16
+                       Test.Cardano.Prelude.Gen
+                       Test.Cardano.Prelude.Golden
+                       Test.Cardano.Prelude.Orphans
+                       Test.Cardano.Prelude.QuickCheck.Arbitrary
+                       Test.Cardano.Prelude.QuickCheck.Property
+                       Test.Cardano.Prelude.Tripping
 
   build-depends:       base
                      , aeson
@@ -35,10 +43,14 @@ library
                      , QuickCheck
                      , quickcheck-instances
                      , text
-                     , time-units
 
   default-language:    Haskell2010
   default-extensions:  NoImplicitPrelude
 
-  ghc-options:         -Wall
-                       -Werror
+  ghc-options:         -Weverything
+                       -fno-warn-missing-import-lists
+                       -fno-warn-unsafe
+                       -fno-warn-safe
+
+  if (!flag(development))
+    ghc-options:         -Werror
