--- util/ChangeLog.md
@@ -1,5 +1,0 @@
-# Revision history for cardano-util
-
-## 0.1.0.0 -- 2018-09-21
-
-* First version. Released on an unsuspecting world.
