--- util/README.md
@@ -1,3 +1,0 @@
-# cardano-util
-
-Utility types, functions, and instances for the Cardano project
