--- util/stack.yaml
@@ -1,8 +1,0 @@
-resolver: ../snapshot.yaml
-
-packages:
-  - ./
-  - test/
-
-extra-deps:
-  - ../prelude
